Added chunk rendering, cleanly bounded light distance

+++ src/chunk.rs
@@ -5,11 +5,11 @@ use crate::light::Light;
 use crate::tile::Tile;
 use crate::vector2::Vector2;
 
-pub const CHUNK_SIZE: usize = 16;
+pub const CHUNK_SIZE: u32 = 16;
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
 pub struct Chunk {
-    pub layout: [[Tile; CHUNK_SIZE]; CHUNK_SIZE],
+    pub layout: [[Tile; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
     pub lights: Vec<(Light, Vector2)>,
 }
